Split multipart into separate modules
